+++ examples/stencil-components/angular/src/app/app.component.ts
@@ -3,17 +3,21 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  // styleUrls: ['./app.component.css']
 })
 export class AppComponent {
   progressValue = 10;
   checkboxChecked = false;
   checkboxError = false;
   checkboxDisabled = false;
+  switchChecked = false;
+  switchDisabled = false;
   radioError = false;
   radioDisabled = false;
   radioChecked = false;
   textFieldValue = '';
+  searchFieldValue = '';
+  searchBarValue = '';
   radioButtonValue = false;
   numberIndicator = 1;
 
@@ -45,6 +49,13 @@ export class AppComponent {
     this.checkboxChecked = !this.checkboxChecked;
   }
 
+  toggleSwitchValue() {
+    console.log("switch value change")
+    this.switchChecked = !this.switchChecked;
+  }
+  toggleSwitchDisabled() {
+    this.switchDisabled = !this.switchDisabled;
+  }
   toggleRadioBtnDisabled() {
     this.radioDisabled = !this.radioDisabled;
   }
@@ -57,9 +68,22 @@ export class AppComponent {
   }
 
 
-  handleIfxInput(event: CustomEvent) {
+  handleTextInput(event: CustomEvent) {
     console.log("textInput value change")
     this.textFieldValue = event.detail;
+
+  }
+
+
+  handleSearchField(event: CustomEvent) {
+    console.log("searchField value change: ", event.detail)
+    this.searchFieldValue = event.detail;
+
+  }
+
+  handleSearchBar(event: CustomEvent) {
+    console.log("searchBar value change: ", event.detail)
+    this.searchBarValue = event.detail;
 
   }