+++ examples/wrapper-components/react-javascript/src/App.js
@@ -11,27 +11,42 @@ import Sidebar from './components/Sidebar/Sidebar'
 import NumberIndicator from './components/NumberIndicator/NumberIndicator'
 import Spinner from './components/Spinner/Spinner'
 import Checkbox from './components/Checkbox/Checkbox';
-import Tag from './components/Tag/Tag';
+import Navbar from './components/Navbar/Navbar';
+import SearchField from './components/SearchField/SearchField';
+import IconButton from './components/IconButton/IconButton';
 import Tab from './components/Tab/Tab';
+import Tag from './components/Tag/Tag';
+import Switch from './components/Switch/Switch';
+import Card from './components/Card/Card';
+
 function App() {
   return (
     <div>
+      <Navbar />
       <h1>Stencil Framework integration - React + JS</h1>
-
-      <h2>Tag</h2>
-      <Tag>Label Tag</Tag>
-      <br />
-
-      <h2>Search Bar</h2>
-      <SearchBar />
-      <br />
 
       <h2>Alert</h2>
       <Alert />
       <br />
 
+      <h2>IconButton</h2>
+      <IconButton />
+      <br />
+
+      <h2>Tab</h2>
+      <Tab />
+      <br />
+
+      <h2>Tag</h2>
+      <Tag />
+      <br />
+
       <h2>Button</h2>
       <Button />
+      <br />
+
+      <h2>Card</h2>
+      <Card />
       <br />
 
       <h2>Progress Bar</h2>
@@ -56,21 +71,35 @@ function App() {
 
       <h2>Link</h2>
       <Link />
+      <br />
 
-      <h2>Tab</h2>
-      <Tab />
+      <h2>Spinner</h2>
+      <Link />
+      <br />
+
+      <h2>Search Bar</h2>
+      <SearchBar />
+      <br />
+
+      <h2>Search Field</h2>
+      <SearchField />
 
       <h2>Spinner</h2>
       <Spinner />
       <br />
 
-      <h2>Number Indicator</h2>
+      <h2>Number indicator</h2>
       <NumberIndicator />
       <br />
 
       <h2>Sidebar</h2>
       <Sidebar />
       <br />
+
+      <h2>Switch</h2>
+      <Switch />
+      <br />
+
     </div>
 
   )