+++ examples/wrapper-components/vue-typescript/src/App.vue
@@ -1,14 +1,19 @@
 
 
 <template>
-  <h1 class="header">Stencil Framework integration - Vue + TS </h1>
+  <Navbar />
+  <h1>Stencil Framework integration - Vue + TS </h1>
   <Alert />
   <Button />
   <Checkbox />
   <RadioButton />
   <ProgressBar />
-  <Search />
+
+  <SearchField />
+  <SearchBar />
   <TextInput />
+  <Search />
+  <TextField />
   <Tag />
   <Link />
   <NumberIndicator />
@@ -16,28 +21,33 @@
   <Spinner />
   <Sidebar />
   <Tabs />
+  <IconButton />
+  <Card />
 </template>
 
 
 <script lang="ts" setup>
+
+
 import Alert from './components/Alert.vue'
 import Button from './components/Button.vue'
 import Checkbox from './components/Checkbox.vue'
 import NumberIndicator from './components/NumberIndicator.vue'
 import ProgressBar from './components/ProgressBar.vue'
-import Search from './components/Search.vue'
+import SearchField from './components/SearchField.vue'
+import SearchBar from './components/SearchBar.vue'
 import Sidebar from './components/Sidebar.vue'
-import TextInput from './components/TextInput.vue'
+import TextField from './components/TextField.vue'
 import Link from './components/Link.vue';
 import Spinner from './components/Spinner.vue';
+import IconButton from './components/IconButton.vue'
 import Tabs from './components/Tabs.vue'
 import RadioButton from './components/RadioButton.vue'
 import Tag from './components/Tag.vue'
+import Navbar from './components/Navbar.vue'
+import Card from './components/Card.vue'
+
 
 </script>
 
-<style scoped>
-.header {
-  text-align: center;
-}
-</style>
+