+++ packages/components/CHANGELOG.md
@@ -1,3 +1,249 @@
+# v19.1.0 (Tue Aug 01 2023)
+
+#### 🚀 Enhancement
+
+- 188 review search input [#219](https://github.com/Infineon/infineon-design-system-stencil/pull/219) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### 🐛 Bug Fix
+
+- added height 100% to sidebar host [#335](https://github.com/Infineon/infineon-design-system-stencil/pull/335) ([@tishoyanchev](https://github.com/tishoyanchev))
+
+#### 🔩 Dependency Updates
+
+- Bump semver from 5.7.1 to 5.7.2 in /examples/wrapper-components/vue-typescript [#342](https://github.com/Infineon/infineon-design-system-stencil/pull/342) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+#### Authors: 3
+
+- [@dependabot[bot]](https://github.com/dependabot[bot])
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v19.0.0 (Tue Aug 01 2023)
+
+#### 💥 Breaking Change
+
+- 165 rework card component resolve overflow icon and image boolean issues implement new structure [#201](https://github.com/Infineon/infineon-design-system-stencil/pull/201) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 2
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.7.1 (Tue Aug 01 2023)
+
+#### 🐛 Bug Fix
+
+- updated navbar [#330](https://github.com/Infineon/infineon-design-system-stencil/pull/330) ([@tishoyanchev](https://github.com/tishoyanchev))
+
+#### Authors: 1
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+
+---
+
+# v18.7.0 (Thu Jul 27 2023)
+
+#### 🚀 Enhancement
+
+- 233 review switch [#309](https://github.com/Infineon/infineon-design-system-stencil/pull/309) (lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 2
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.6.0 (Tue Jul 25 2023)
+
+#### 🚀 Enhancement
+
+- added icon button doc [#302](https://github.com/Infineon/infineon-design-system-stencil/pull/302) ([@SergejLust](https://github.com/SergejLust) lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx) [@github-actions[bot]](https://github.com/github-actions[bot]))
+
+#### Authors: 4
+
+- [@github-actions[bot]](https://github.com/github-actions[bot])
+- [@SergejLust](https://github.com/SergejLust)
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.5.6 (Tue Jul 25 2023)
+
+#### 🐛 Bug Fix
+
+- 306 test wrapper components with vite [#307](https://github.com/Infineon/infineon-design-system-stencil/pull/307) (lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 2
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.5.5 (Mon Jul 24 2023)
+
+#### 🐛 Bug Fix
+
+- added text field doc [#298](https://github.com/Infineon/infineon-design-system-stencil/pull/298) ([@SergejLust](https://github.com/SergejLust) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 2
+
+- [@SergejLust](https://github.com/SergejLust)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.5.4 (Mon Jul 24 2023)
+
+#### 🐛 Bug Fix
+
+- 251 fix icon button styling [#268](https://github.com/Infineon/infineon-design-system-stencil/pull/268) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### 🔩 Dependency Updates
+
+- Bump semver from 5.7.1 to 5.7.2 in /examples/stencil-components/angular [#295](https://github.com/Infineon/infineon-design-system-stencil/pull/295) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump semver from 5.7.1 to 5.7.2 in /examples/wrapper-components/react-typescript [#296](https://github.com/Infineon/infineon-design-system-stencil/pull/296) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+#### Authors: 3
+
+- [@dependabot[bot]](https://github.com/dependabot[bot])
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.5.3 (Fri Jul 21 2023)
+
+#### 🐛 Bug Fix
+
+- deploying preview urls for the example apps using wrapper components [#272](https://github.com/Infineon/infineon-design-system-stencil/pull/272) (lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 2
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.5.2 (Fri Jul 21 2023)
+
+#### 🐛 Bug Fix
+
+- added badge doc [#294](https://github.com/Infineon/infineon-design-system-stencil/pull/294) ([@SergejLust](https://github.com/SergejLust) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 2
+
+- [@SergejLust](https://github.com/SergejLust)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.5.1 (Fri Jul 21 2023)
+
+#### 🐛 Bug Fix
+
+- Documentation checkbox [#285](https://github.com/Infineon/infineon-design-system-stencil/pull/285) ([@SergejLust](https://github.com/SergejLust) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 2
+
+- [@SergejLust](https://github.com/SergejLust)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.5.0 (Fri Jul 21 2023)
+
+#### 🚀 Enhancement
+
+- Documentation: Range [#282](https://github.com/Infineon/infineon-design-system-stencil/pull/282) ([@SergejLust](https://github.com/SergejLust) [@dependabot[bot]](https://github.com/dependabot[bot]) [@verena-ifx](https://github.com/verena-ifx) [@github-actions[bot]](https://github.com/github-actions[bot]) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 5
+
+- [@dependabot[bot]](https://github.com/dependabot[bot])
+- [@github-actions[bot]](https://github.com/github-actions[bot])
+- [@SergejLust](https://github.com/SergejLust)
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.4.1 (Fri Jul 21 2023)
+
+#### 🐛 Bug Fix
+
+- 252 fix navbar component position inconsistency [#256](https://github.com/Infineon/infineon-design-system-stencil/pull/256) ([@tishoyanchev](https://github.com/tishoyanchev) lechneve@ISCN5CG1201S7M.infineon.com)
+
+#### Authors: 2
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v18.4.0 (Fri Jul 21 2023)
+
+#### 🚀 Enhancement
+
+- Documentation sidebar [#284](https://github.com/Infineon/infineon-design-system-stencil/pull/284) ([@SergejLust](https://github.com/SergejLust))
+- 265 documentation documentation spinner [#286](https://github.com/Infineon/infineon-design-system-stencil/pull/286) ([@SergejLust](https://github.com/SergejLust))
+- added accordion documentation [#290](https://github.com/Infineon/infineon-design-system-stencil/pull/290) ([@SergejLust](https://github.com/SergejLust))
+
+#### Authors: 1
+
+- [@SergejLust](https://github.com/SergejLust)
+
+---
+
+# v18.3.3 (Thu Jul 20 2023)
+
+#### ⚠️ Pushed to `master`
+
+- Delete commit_changes.yml ([@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 1
+
+- [@verena-ifx](https://github.com/verena-ifx)
+
+---
+
+# v18.3.2 (Thu Jul 20 2023)
+
+#### ⚠️ Pushed to `master`
+
+- Create commit_changes.yml ([@verena-ifx](https://github.com/verena-ifx))
+
+#### 🔩 Dependency Updates
+
+- Bump word-wrap from 1.2.3 to 1.2.4 in /examples/wrapper-components/react-javascript [#279](https://github.com/Infineon/infineon-design-system-stencil/pull/279) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump word-wrap from 1.2.3 to 1.2.4 in /examples/wrapper-components/react-typescript [#278](https://github.com/Infineon/infineon-design-system-stencil/pull/278) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump word-wrap from 1.2.3 to 1.2.4 in /examples/stencil-components/react [#280](https://github.com/Infineon/infineon-design-system-stencil/pull/280) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump word-wrap from 1.2.3 to 1.2.4 in /packages/components [#281](https://github.com/Infineon/infineon-design-system-stencil/pull/281) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+#### Authors: 2
+
+- [@dependabot[bot]](https://github.com/dependabot[bot])
+- [@verena-ifx](https://github.com/verena-ifx)
+
+---
+
+# v18.3.1 (Wed Jul 19 2023)
+
+#### 🐛 Bug Fix
+
+- 271 documentation radio button [#276](https://github.com/Infineon/infineon-design-system-stencil/pull/276) ([@SergejLust](https://github.com/SergejLust))
+
+#### Authors: 1
+
+- [@SergejLust](https://github.com/SergejLust)
+
+---
+
 # v18.3.0 (Mon Jul 17 2023)
 
 #### 🚀 Enhancement